Retornando ao ponto inicial do erro
