+++ inc/colors.h
@@ -1,0 +1,22 @@
+/*******************************************************************************
+ * Cores
+ ******************************************************************************/
+
+enum colors
+{
+    red,
+    green,
+    blue,
+    yellow,
+    purple,
+    white,
+    black
+};
+
+static const char color_red[] = "vermelho";
+static const char color_green[] = "verde";
+static const char color_blue[] = "azul";
+static const char color_yellow[] = "amarelo";
+static const char color_purple[] = "roxo";
+static const char color_white[] = "branco";
+static const char color_black[] = "preto";
